+++ CHANGELOG.md
@@ -2,10 +2,9 @@
 
 ## Development
 
-*   Add installation support for Python 3.8 and 3.9 ([#135](https://github.com/AI-SDC/ACRO/pull/135))
-
 Changes:
-*   Add a directory of outputs to an acro object and creates a results file for checking.
+*   Add installation support for Python 3.8 and 3.9 ([#135](https://github.com/AI-SDC/ACRO/pull/135))
+*   Add a directory of outputs to an acro object and creates a results file for checking ([#130](https://github.com/AI-SDC/ACRO/pull/130))
 
 ## Version 0.4.2 (Jul 13, 2023)