+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes:
 *   Improve screen formatting of tables ([#139](https://github.com/AI-SDC/ACRO/pull/139))
 *   Integrate Stata support ([#142](https://github.com/AI-SDC/ACRO/pull/142))
 *   Fix crosstab when index or columns is a list and there is an aggregation function ([#147](https://github.com/AI-SDC/ACRO/pull/147))
+*   Add support for survival analysis ([#145](https://github.com/AI-SDC/ACRO/pull/145))
 
 ## Version 0.4.2 (Jul 13, 2023)