+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes:
 *   Add support for survival analysis ([#145](https://github.com/AI-SDC/ACRO/pull/145))
 *   Update pandas version dependency to 1.5 ([#150](https://github.com/AI-SDC/ACRO/pull/150))
 *   Refactor ACRO class ([#152](https://github.com/AI-SDC/ACRO/pull/152))
+*   Adding support for table function in R ([#153](https://github.com/AI-SDC/ACRO/pull/153))
 
 ## Version 0.4.2 (Jul 13, 2023)